Image controller

+++ src/main/java/com/technicaltest/productservice/facade/ProductFacade.java
@@ -3,6 +3,7 @@ package com.technicaltest.productservice.facade;
 import com.technicaltest.productservice.dto.ProductData;
 import com.technicaltest.productservice.entity.ProductEntity;
 import com.technicaltest.productservice.mapper.ProductMapper;
+import com.technicaltest.productservice.service.ImageService;
 import com.technicaltest.productservice.service.ProductService;
 import com.technicaltest.productservice.utils.Utils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +16,9 @@ public class ProductFacade {
 
     @Autowired
     private ProductService productService;
+
+    @Autowired
+    private ImageService imageService;
 
     @Autowired
     private ProductMapper productMapper;
@@ -43,9 +47,11 @@ public class ProductFacade {
         if(Utils.isEmpty(existing.getId())){
             throw new Exception("Product not found");
         }
+        if(!Utils.isEmpty(existing.getImage()) && existing.getImage() != productData.getImage()){
+            imageService.delete(existing.getImage());
+        }
         ProductEntity requestBody = productMapper.productDataToProductEntity(productData);
         requestBody.setId(existing.getId());
-
         ProductEntity product = productService.save(requestBody);
         return productMapper.productEntityToProductData(product);
     }
@@ -60,6 +66,12 @@ public class ProductFacade {
     }
 
     public void delete(Long productId) {
-        productService.delete(productId);
+        ProductEntity existing = productService.getById(productId);
+        if(!Utils.isEmpty(existing)){
+            if(!Utils.isEmpty(existing.getImage())){
+                imageService.delete(existing.getImage());
+            }
+            productService.delete(productId);
+        }
     }
 }
